Move the single transaction operations
