+++ src/datastruct.rs
@@ -1,30 +1,6 @@
 use chrono::Utc;
 use serde_derive::{Deserialize, Serialize};
 
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Copy, Clone)]
-pub enum EntryType {
-    Debit,
-    Credit,
-}
-
-impl EntryType {
-    pub fn from_i32(value: i32) -> EntryType {
-        match value {
-            0 => EntryType::Credit,
-            1 => EntryType::Debit,
-            _ => panic!("Unknown value: {}", value),
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct NewTransaction {
-    pub name: String,
-    pub balance: i32,
-    pub from: i32,
-    pub to: i32,
-}
-
 #[derive(Debug, Serialize, Deserialize)]
 pub struct IdRequest {
     pub id: i32,
@@ -44,15 +20,6 @@ pub struct Transaction {
 }
 
 #[derive(Debug, Serialize)]
-pub struct Entry {
-    pub id: i32,
-    pub account: i32,
-    pub transaction_id: i32,
-    pub balance: i32,
-    pub entry_type: EntryType,
-}
-
-#[derive(Debug, Serialize)]
 pub struct SqlResult {
     pub value: i32,
 }
@@ -63,18 +30,4 @@ pub struct Currency {
     pub numeric_code: i32,
     pub minor_unit: i32,
     pub name: String,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn entry_correctly_types() {
-        let debit = EntryType::from_i32(1);
-        let credit = EntryType::from_i32(0);
-
-        assert_eq!(debit, EntryType::Debit);
-        assert_eq!(credit, EntryType::Credit)
-    }
 }
